feat: 잘못된 Tag 값에 대한 진단

+++ TaggedUnion.Core/TaggedUnionCaseAttribute.cs
@@ -11,11 +11,11 @@ public sealed class TaggedUnionCaseAttribute : Attribute
 
     public string? ParamName { get; }
 
-    public int Tag { get; }
+    public byte Tag { get; }
     #endregion
 
     #region Constructors
-    public TaggedUnionCaseAttribute(Type type, string? paramName = null, int tag = 0)
+    public TaggedUnionCaseAttribute(Type type, string? paramName = null, byte tag = 0)
     {
         Type = type;
         ParamName = paramName;
