+++ TaggedUnion.Generator/UnionCaseContext.cs
@@ -4,5 +4,5 @@ internal sealed record UnionCaseContext(
     UnionCaseStorageKind StorageKind,
     string FullyQualifiedTypeName,
     string ParamName,
-    int Tag
+    byte Tag
 );
